feat: 验证码判定逻辑修改

+++ router_handler/login.js
@@ -48,40 +48,45 @@ const sendAuthCode = async (req, res) => {
 	}
 }
 
-const codeLogin = async (req, res) => {
-	const { account, authCode, type = '' } = req.body
-	if (!account || !authCode) return res.sendError(400, 'account or authCode is required')
+const judgeAuthCode = async (req, res) => {
 	try {
-		const existingUser = await User.findOne({ email: account })
+		const { account, authCode } = req.body
+		if (!account || !authCode) return res.sendError(400, 'account or authCode is required')
 		const authCodeDoc = await AuthCode.findOne({ email: account })
 		if (!authCodeDoc || authCode !== authCodeDoc.code || Date.now() > authCodeDoc.expires)
+			return res.sendError(400, 'Expired authCode')
+		if (authCode !== authCodeDoc.code) return res.sendError(400, 'Invalid authCode')
+		await AuthCode.deleteMany({ email: account })
+		return res.sendSuccess('Correct authCode')
+	} catch (error) {
+		console.error('Error in authCode:', error)
+	}
+}
+
+const codeLogin = async (req, res) => {
+	const { account, authCode } = req.body
+	if (!account || !authCode) return res.sendError(400, 'account or authCode is required')
+	try {
+		const [existingUser, authCodeDoc] = await Promise.all([
+			User.findOne({ email: account }),
+			AuthCode.findOne({ email: account })
+		])
+		if (!authCodeDoc || authCode !== authCodeDoc.code || Date.now() > authCodeDoc.expires)
 			return res.sendError(400, 'Invalid or expired auth code')
-		let user
-		let token
-		if (existingUser) {
-			user = existingUser
-			user.lastLoginDate = new Date().toISOString().split('T')[0]
-		} else {
-			user = new User({
+		const user =
+			existingUser ||
+			new User({
 				email: account,
 				password: bcrypt.hashSync(authCode, 10),
 				nickname: account,
 				sex: 2,
 				birthDate: new Date().toISOString().split('T')[0],
-				registrationDate: new Date().toISOString().split('T')[0],
-				lastLoginDate: new Date().toISOString().split('T')[0]
+				registrationDate: new Date()
 			})
-			token = generateToken(user)
-			await user.save()
-		}
 		user.lastLoginDate = new Date()
-		if (type !== 'auth') {
-			user.token = token
-		}
-		token = generateToken(user)
-		await user.save()
-		await AuthCode.deleteMany({ email: account })
-		return res.sendSuccess({ token }, 'Login/Register success')
+		user.token = generateToken(user)
+		await Promise.all([user.save(), AuthCode.deleteMany({ email: account })])
+		return res.sendSuccess({ token: user.token }, 'Register/Login success')
 	} catch (error) {
 		console.error(error)
 		return res.sendError(500, 'Server error')
@@ -107,19 +112,19 @@ const logout = async (req, res) => {
 	const { account } = req.body
 	if (!account) return res.sendError(400, 'account is required')
 	try {
-		const user = await User.findOne({ email: account })
-		if (!user) return res.sendError(404, 'User not found')
-		user.token = null
-		await user.save()
-		res.sendSuccess({ message: 'Logout successful' })
+		const result = await User.updateOne({ email: account }, { $set: { token: null } })
+		if (result.matchedCount === 0) return res.sendError(404, 'User not found')
+		if (result.modifiedCount === 0) return res.sendSuccess({ message: 'User already logged out' })
+		return res.sendSuccess({ message: 'Logout successful' })
 	} catch (error) {
 		console.error('Error in logout:', error)
-		res.sendError(500, 'Internal server error')
+		return res.sendError(500, 'Internal server error')
 	}
 }
 
 module.exports = {
 	sendAuthCode,
+	judgeAuthCode,
 	codeLogin,
 	passwordLogin,
 	logout
